Refactor landing plugin
